refactor: update ops references in auto_kms
