+++ pkgs/standards/autoapi/autoapi/v3/bindings/rpc.py
@@ -17,8 +17,8 @@ from typing import (
 
 from pydantic import BaseModel
 
-from ..opspec import OpSpec
-from ..opspec.types import PHASES
+from ..ops import OpSpec
+from ..ops.types import PHASES
 from ..runtime import executor as _executor  # expects _invoke(request, db, phases, ctx)
 
 # Prefer Kernel phase-chains if available (atoms + system steps + hooks)
