+++ pkgs/standards/autoapi/autoapi/v3/ops/types.py
@@ -1,4 +1,4 @@
-# autoapi/v3/opspec/types.py
+# autoapi/v3/ops/types.py
 from __future__ import annotations
 
 from dataclasses import dataclass, field
