+++ pkgs/standards/autoapi/autoapi/v3/runtime/kernel.py
@@ -9,7 +9,7 @@ from typing import Any, Callable, Dict, Iterable, List, Mapping, Optional, Seque
 
 from .executor import _invoke, _Ctx
 from . import events as _ev
-from ..opspec.types import PHASES, StepFn
+from ..ops.types import PHASES, StepFn
 
 logger = logging.getLogger(__name__)
