+++ pkgs/standards/autoapi/tests/i9n/test_bindings_modules.py
@@ -20,7 +20,7 @@ from autoapi.v3.bindings import (
     schemas as schemas_binding,
 )
 from autoapi.v3.specs import ColumnSpec, F, IO, S
-from autoapi.v3.opspec import resolve
+from autoapi.v3.ops import resolve
 from autoapi.v3.runtime import executor as _executor
 from autoapi.v3.specs import shortcuts as sc
