+++ pkgs/standards/autoapi/tests/i9n/test_bulk_docs_client.py
@@ -2,7 +2,7 @@ import pytest
 from httpx import AsyncClient, ASGITransport
 
 from autoapi.v3.bindings.rest import _build_router
-from autoapi.v3.opspec import OpSpec
+from autoapi.v3.ops import OpSpec
 from autoapi.v3.orm.tables import Base
 from autoapi.v3.orm.mixins import GUIDPk, BulkCapable
 from autoapi.v3.types import App, Column, String
