+++ pkgs/standards/autoapi/tests/i9n/test_opspec_effects_i9n_test.py
@@ -8,7 +8,7 @@ from sqlalchemy.orm import sessionmaker
 from autoapi.v3 import core as _core
 from autoapi.v3.bindings.model import bind
 from autoapi.v3.decorators import hook_ctx
-from autoapi.v3.opspec.types import PHASES
+from autoapi.v3.ops.types import PHASES
 from autoapi.v3.runtime import system as runtime_system
 from autoapi.v3.runtime.kernel import build_phase_chains
 from autoapi.v3.specs import IO, S, acol
