+++ pkgs/standards/autoapi/tests/unit/test_alias_ctx_op_alias_attributes.py
@@ -1,7 +1,7 @@
 from pydantic import BaseModel
 
 from autoapi.v3 import alias_ctx, alias, schema_ctx
-from autoapi.v3.opspec import resolve
+from autoapi.v3.ops import resolve
 from autoapi.v3.bindings import build_schemas, build_handlers
