+++ pkgs/standards/autoapi/tests/unit/test_op_alias.py
@@ -1,7 +1,7 @@
 from pydantic import BaseModel
 
 from autoapi.v3 import op_alias
-from autoapi.v3.opspec import resolve
+from autoapi.v3.ops import resolve
 
 
 def _get_spec(**decorator_kwargs):
