+++ pkgs/standards/autoapi/tests/unit/test_opspec_effects.py
@@ -8,7 +8,7 @@ from sqlalchemy import String, create_engine
 from sqlalchemy.orm import sessionmaker
 
 from autoapi.v3.bindings.model import bind
-from autoapi.v3.opspec.types import PHASES
+from autoapi.v3.ops.types import PHASES
 from autoapi.v3.runtime.kernel import build_phase_chains
 from autoapi.v3.runtime import system as runtime_system
 from autoapi.v3.specs import IO, S, acol
