+++ pkgs/standards/autoapi/tests/unit/test_request_body_schema.py
@@ -1,7 +1,7 @@
 from autoapi.v3.types import App
 
 from autoapi.v3.bindings.rest import _build_router
-from autoapi.v3.opspec import OpSpec
+from autoapi.v3.ops import OpSpec
 from autoapi.v3.orm.tables import Base
 from autoapi.v3.orm.mixins import GUIDPk
 from autoapi.v3.types import Column, String
