+++ pkgs/standards/autoapi/tests/unit/test_request_response_examples.py
@@ -1,5 +1,5 @@
 from autoapi.v3.bindings.rest import _build_router
-from autoapi.v3.opspec import OpSpec
+from autoapi.v3.ops import OpSpec
 from autoapi.v3.orm.tables import Base
 from autoapi.v3.orm.mixins import GUIDPk, BulkCapable, Mergeable
 from autoapi.v3.types import Column, String, App
