+++ pkgs/standards/autoapi/tests/unit/test_rest_bulk_delete_suppresses_clear.py
@@ -1,6 +1,6 @@
 from autoapi.v3.bindings.rest import build_router_and_attach
 from autoapi.v3.orm.mixins import GUIDPk
-from autoapi.v3.opspec import OpSpec
+from autoapi.v3.ops import OpSpec
 from autoapi.v3.orm.tables import Base
 from autoapi.v3.types import Column, String
