+++ pkgs/standards/autoapi/tests/unit/test_rest_rpc_parity_default_ops.py
@@ -2,7 +2,7 @@ import pytest
 
 from autoapi.v3.autoapi import AutoAPI
 from autoapi.v3.orm.mixins import GUIDPk, BulkCapable, Mergeable
-from autoapi.v3.opspec.types import CANON
+from autoapi.v3.ops.types import CANON
 from autoapi.v3.orm.tables import Base
 from autoapi.v3.types import Column, String
