+++ pkgs/standards/autoapi/tests/unit/test_v3_schemas_and_decorators.py
@@ -7,7 +7,7 @@ from autoapi.v3 import (
     schema_ctx,
     SchemaRef,
 )
-from autoapi.v3.opspec import resolve
+from autoapi.v3.ops import resolve
 from autoapi.v3.decorators import collect_decorated_ops
 from autoapi.v3.bindings import build_schemas, build_hooks, build_handlers, build_rest
